The local master had been sitting at 2026-08-07 with 114 unpushed commits while
origin moved on 741 commits to 2026-09-03. All 114 are Tracker bookkeeping for
the 2026-08-05 full audit -- none of their subjects appears in origin/master, so
none could be dropped; the audit's code fixes reached origin through the
full-audit-fixes-2026-08-05 branch and only the disposition ledger stayed behind.

One conflict, add/add on audits/full_audit_2026-08-05/_TRACKER.md, resolved to
origin's version. The two differ in exactly one way: origin's links point into
implemented/, where 32 of the audit files have since been moved, and the local
copy still links to the flat paths they no longer occupy. Row count and
dispositions are identical on both sides (338 rows; 335 RESOLVED, 4 FUTURE,
3 REJECTED, 1 DOC), verified after the resolution rather than assumed from the
conflict being small, and the 32 links now match the 32 files on disk.

src/, tests/ and pyproject.toml are byte-identical to origin/master after the
merge, checked rather than inferred: the 114 local commits touched only audits/.

+++ src/mlframe/feature_selection/filters/_permutation_null_resident.py
@@ -168,7 +168,14 @@ def gen_target_shuffles_cupy(y_codes: np.ndarray, nperm: int, dtype: type, rando
         seed = 0x9E3779B9 if random_seed is None else (int(random_seed) & 0x7FFFFFFF)
         _rng = cp.random.default_rng(seed)
         d_y = cp.asarray(np.ascontiguousarray(y_codes).astype(dtype, copy=False))  # one tiny (n,) upload
-        keys = _rng.random((nperm, n), dtype=cp.float32)  # (nperm, n) i.i.d. sort keys, on device
+        # float64 keys, not float32. float32 uniforms occupy only ~1.68e7 grid points, so the expected number of
+        # TIED pairs per row is n^2 / 2^25 -- about 10,700 at n=600k and 1.2e5 at n=2M, and this path is gated to
+        # fire at exactly those sizes. `argsort` breaks ties by index, so tied positions keep their original
+        # relative order and each row carries a small positive correlation with the identity permutation instead
+        # of being a uniform draw. The bias is conservative (the null band inflates, so the maxT floor is slightly
+        # too strict and rejects marginal true candidates), but it makes the GPU floor a different estimator from
+        # the CPU Fisher-Yates one, which the docstring's "statistically equivalent" claim does not cover.
+        keys = _rng.random((nperm, n), dtype=cp.float64)  # (nperm, n) i.i.d. sort keys, on device
         order = cp.argsort(keys, axis=1)  # per-row permutation indices
         del keys
         out = d_y[order]  # (nperm, n) gathered shuffles, on device
